Moved runtime leak detection to utils.py.
