+++ open_instruct/grpo_fast.py
@@ -56,7 +56,6 @@ import traceback
 from argparse import Namespace
 from collections import defaultdict
 from dataclasses import asdict, dataclass, field
-from multiprocessing import resource_tracker as _rt
 from queue import Empty, Queue
 from typing import Any, Callable, Dict, Iterable, Iterator, List, Literal, Optional, Union
 
@@ -146,132 +145,6 @@ api = HfApi()
 INVALID_LOGPROB = 1.0
 
 
-# ---- Runtime leak detection -----------------------------------------------------------------
-
-DEFAULT_THREAD_ALLOWLIST = {
-    "MainThread",
-    "pytest-watcher",  # pytest
-    "pydevd.",  # debugger
-    "IPythonHistorySavingThread",
-    "raylet_client",  # ray internal when still up during test body
-}
-
-DEFAULT_THREAD_ALLOW_PREFIXES = {
-    "ThreadPoolExecutor-",  # executors create transient threads; adjust if you join them
-    "ray-",  # ray internal threads
-    "grpc-default-executor",  # grpc internal
-}
-
-
-@dataclass
-class LeakReport:
-    bad_threads: List[threading.Thread] = field(default_factory=list)
-    bad_processes: List[mp.Process] = field(default_factory=list)
-    ray_actors: list = field(default_factory=list)
-    ray_tasks: list = field(default_factory=list)
-    ray_workers: list = field(default_factory=list)
-    leaked_semaphores: List[str] = field(default_factory=list)
-    leaked_shm: List[str] = field(default_factory=list)
-
-    @property
-    def is_clean(self) -> bool:
-        return not (
-            self.bad_threads
-            or self.bad_processes
-            or self.ray_actors
-            or self.ray_tasks
-            or self.ray_workers
-            or self.leaked_semaphores
-            or self.leaked_shm
-        )
-
-    def pretty(self) -> str:
-        lines = []
-        if self.bad_threads:
-            lines.append("Leaked threads:")
-            for t in self.bad_threads:
-                target = getattr(t, "_target", None)
-                tgt_name = getattr(target, "__name__", repr(target)) if target else "?"
-                lines.append(f"  - {t.name} (alive={t.is_alive()}, daemon={t.daemon}, target={tgt_name})")
-        if self.bad_processes:
-            lines.append("Leaked multiprocessing children:")
-            for p in self.bad_processes:
-                lines.append(f"  - PID {p.pid} alive={p.is_alive()} name={p.name}")
-        if self.ray_actors:
-            lines.append("Live Ray actors:")
-            for a in self.ray_actors:
-                lines.append(f"  - {a.get('class_name')} id={a.get('actor_id')}")
-        if self.ray_tasks:
-            lines.append("Live Ray tasks:")
-            for t in self.ray_tasks:
-                lines.append(f"  - {t.get('name')} id={t.get('task_id')}")
-        if self.ray_workers:
-            lines.append("Live Ray workers:")
-            for w in self.ray_workers:
-                lines.append(f"  - pid={w.get('pid')} id={w.get('worker_id')}")
-        if self.leaked_semaphores:
-            lines.append("Leaked POSIX semaphores (multiprocessing):")
-            for name in self.leaked_semaphores:
-                lines.append(f"  - {name}")
-        if self.leaked_shm:
-            lines.append("Leaked shared_memory blocks:")
-            for name in self.leaked_shm:
-                lines.append(f"  - {name}")
-        return "\n".join(lines) if lines else "No leaks detected."
-
-
-def check_runtime_leaks(
-    thread_allowlist: Iterable[str] = DEFAULT_THREAD_ALLOWLIST,
-    thread_allow_prefixes: Iterable[str] = DEFAULT_THREAD_ALLOW_PREFIXES,
-    include_daemon_threads: bool = False,
-) -> LeakReport:
-    """
-    Inspect runtime state for leftovers.
-
-    Returns a LeakReport; call .is_clean or .pretty().
-    """
-    report = LeakReport()
-
-    # Threads
-    for t in threading.enumerate():
-        if t.name in thread_allowlist or any(t.name.startswith(p) for p in thread_allow_prefixes):
-            continue
-        if not include_daemon_threads and t.daemon:
-            continue
-        if t is threading.main_thread():
-            continue
-        # ignore ones already finished
-        if not t.is_alive():
-            continue
-        report.bad_threads.append(t)
-
-    # Multiprocessing children
-    for child in mp.active_children():
-        if child.is_alive():
-            report.bad_processes.append(child)
-
-    # Ray state (only if ray is present & initialized)
-    try:
-        import ray
-        from ray.util import state as ray_state
-
-        if ray_state is not None and ray is not None and ray.is_initialized():
-            report.ray_actors = ray_state.list_actors(filters=[("state", "=", "ALIVE")])
-            report.ray_tasks = ray_state.list_tasks(filters=[("state", "=", "RUNNING")])
-            report.ray_workers = ray_state.list_workers(filters=[("is_alive", "=", True)])
-    except Exception:
-        pass
-
-    # Multiprocessing resource_tracker cache (private API, so guard carefully)
-    if _rt is not None and hasattr(_rt, "_resource_tracker"):
-        cache = getattr(_rt._resource_tracker, "_cache", {})
-        for name, (count, rtype) in cache.items():
-            if count > 0 and rtype == "semaphore":
-                report.leaked_semaphores.append(name)
-            if count > 0 and rtype == "shared_memory":
-                report.leaked_shm.append(name)
-
-    return report
 
 
 @dataclass
@@ -2554,7 +2427,8 @@ def main(args: Args, tc: TokenizerConfig, model_config: ModelConfig, num_eval_sa
 
     # Check for runtime leaks before exiting
     logger.info("Checking for runtime leaks...")
-    leak_report = check_runtime_leaks()
+    from open_instruct import utils
+    leak_report = utils.check_runtime_leaks()
     if not leak_report.is_clean:
         logger.warning("Runtime leaks detected:\n" + leak_report.pretty())
     else:
